Return execution plan as a JSON list
HTTP POST requests to /e2data/flink-schedule now return a JSON-formatted
list of job vertices along with the HwResource that has been allocated
for each of them.
The returned list contains only a subset of the original JobGraph
uploaded with the POST request: it contains only those job vertices that
have actually been scheduled to a node (i.e., those which are
"acceleratable" / "offloadable" to an accelerator).
The HwResource is passed along exactly as parsed by Yarn.

Resolves ckatsak/HAIER#1 by checking for absence of an execution plan
right before responding to Flink. In this case, HAIER responds with a
500 INTERNAL SERVER ERROR.

Signed-off-by: Christos Katsakioris <ckatsak@gmail.com>